client-conf: Add allow-autospawn-for-root
Usually autospawning for root is a bad idea, since it can easily
interfere with other users' PulseAudio instances, but in embedded
environments where only root exists, autospawning is fine.

Upstream-Status: Denied [Upstream hasn't so far been convinced that
this feature is needed. The view is that running the graphical session
as root doesn't make sense even on embedded systems, or if it does,
running PulseAudio in the system mode should be a good solution for
that use case. If someone can prove otherwise, the patch would
probably be accepted.]

Imported from:
https://git.openembedded.org/openembedded-core/tree/meta/recipes-multimedia/pulseaudio/pulseaudio/0001-client-conf-Add-allow-autospawn-for-root.patch?h=kirkstone&id=29bf058368e4a6336b74d9aa6880d3176cadb913

Change-Id: Ic566a163a9fdb1c100c65f804323e6d578be075b
Signed-off-by: Tanu Kaskinen <tanu.kaskinen@linux.intel.com>