doxygen/meson.build: remove dependency on doxygen binary
New meson refuses to complete confuguration if doxygen is absent
even if this target is never actually run. As we never run it,
let's remove it, until upstream and/or meson figure out that it's
optional and should be treated as such.

Upstream-Status: Inappropriate [needs proper upstream fix]

Imported from:
https://git.openembedded.org/openembedded-core/tree/meta/recipes-multimedia/pulseaudio/pulseaudio/0001-doxygen-meson.build-remove-dependency-on-doxygen-bin.patch?h=kirkstone&id=29bf058368e4a6336b74d9aa6880d3176cadb913

Change-Id: I690da55acd3ef5131baeb00e22b236f82853671f
Signed-off-by: Alexander Kanavin <alex.kanavin@gmail.com>