Skip moving active sinks if physical sink is empty
:Release Notes:
Implement to Skip moving active sinks if physical sink is empty

:Detailed Notes:
when a new setSoundoutput call is received, the device name is
send to pulseaudio. if the actual sink in that name is not present,
moving the current active stream should be skipped.

:Testing Performed:
Tested on webos build re-based on #351 official binary
Test cases verified: QWA-8969

:QA Notes:
NA

:Issues Addressed:
[PLAT-143568] BT soundoutput audio is not heard

Change-Id: Ia577a1780f2736fd9e7ba4ba0f3e74136629e442